requirements.txt
livekit-agents[openai]~=1.2
python-dotenv
